templates: fix merging SlotBag objects
It's possible for two slot bags to refer to the same function
even though the schemas are different objects. In that case
the merge should be allowed.
